+++ test/data_structures/triangulation.jl
@@ -1475,3 +1475,13 @@ end
         end
     end
 end
+
+@testset "Constructor with only points and triangles" begin
+    points = [(0.0, 0.0), (0.87, 0.0), (1.0006, 0.7766), (0.0, 1.0), (0.5, 0.5)]
+    tri1 = triangulate(points)
+    triangles = tri1.triangles
+    tri2 = DT.Triangulation(points, triangles)
+    @test validate_triangulation(tri2)
+    @test tri1 == tri2
+    @test !DT.has_boundary_nodes(tri2)
+end